+++ src/main/java/com/geosegbar/infra/questionnaire_response/services/QuestionnaireResponseService.java
@@ -3,12 +3,12 @@ package com.geosegbar.infra.questionnaire_response.services;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.geosegbar.entities.QuestionnaireResponseEntity;
 import com.geosegbar.exceptions.NotFoundException;
 import com.geosegbar.infra.questionnaire_response.persistence.jpa.QuestionnaireResponseRepository;
 
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
@@ -21,8 +21,10 @@ public class QuestionnaireResponseService {
 
     @Transactional
     public void deleteById(Long id) {
-        responseRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Resposta do questionário não encontrada para exclusão!"));
+
+        if (!responseRepository.existsById(id)) {
+            throw new NotFoundException("Resposta do questionário não encontrada para exclusão!");
+        }
         responseRepository.deleteById(id);
 
         log.info("QuestionnaireResponse {} deletado.", id);
@@ -30,6 +32,7 @@ public class QuestionnaireResponseService {
 
     @Transactional
     public QuestionnaireResponseEntity save(QuestionnaireResponseEntity response) {
+
         QuestionnaireResponseEntity saved = responseRepository.save(response);
 
         log.info("QuestionnaireResponse {} criado.", saved.getId());
@@ -39,8 +42,10 @@ public class QuestionnaireResponseService {
 
     @Transactional
     public QuestionnaireResponseEntity update(QuestionnaireResponseEntity response) {
-        responseRepository.findById(response.getId())
-                .orElseThrow(() -> new NotFoundException("Resposta do questionário não encontrada para atualização!"));
+        if (!responseRepository.existsById(response.getId())) {
+            throw new NotFoundException("Resposta do questionário não encontrada para atualização!");
+        }
+
         QuestionnaireResponseEntity saved = responseRepository.save(response);
 
         log.info("QuestionnaireResponse {} atualizado.", response.getId());
@@ -48,12 +53,16 @@ public class QuestionnaireResponseService {
         return saved;
     }
 
+    @Transactional(readOnly = true)
     public QuestionnaireResponseEntity findById(Long id) {
+
         return responseRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Resposta do questionário não encontrada!"));
     }
 
+    @Transactional(readOnly = true)
     public List<QuestionnaireResponseEntity> findAll() {
+
         return responseRepository.findAll();
     }
 }
