+++ src/main/java/com/geosegbar/infra/risk_category/services/RiskCategoryService.java
@@ -3,6 +3,7 @@ package com.geosegbar.infra.risk_category.services;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.geosegbar.entities.RiskCategoryEntity;
 import com.geosegbar.exceptions.DuplicateResourceException;
@@ -10,7 +11,6 @@ import com.geosegbar.exceptions.NotFoundException;
 import com.geosegbar.infra.risk_category.persistence.RiskCategoryRepository;
 
 import jakarta.annotation.PostConstruct;
-import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
 @Service
@@ -37,9 +37,9 @@ public class RiskCategoryService {
 
     @Transactional
     public void deleteById(Long id) {
-        riskCategoryRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("Categoria de risco não encontrada para exclusão!"));
-
+        if (!riskCategoryRepository.existsById(id)) {
+            throw new NotFoundException("Categoria de risco não encontrada para exclusão!");
+        }
         riskCategoryRepository.deleteById(id);
     }
 
@@ -54,8 +54,10 @@ public class RiskCategoryService {
 
     @Transactional
     public RiskCategoryEntity update(RiskCategoryEntity riskCategoryEntity) {
-        riskCategoryRepository.findById(riskCategoryEntity.getId())
-                .orElseThrow(() -> new NotFoundException("Categoria de risco não encontrada para atualização!"));
+
+        if (!riskCategoryRepository.existsById(riskCategoryEntity.getId())) {
+            throw new NotFoundException("Categoria de risco não encontrada para atualização!");
+        }
 
         if (riskCategoryRepository.existsByNameAndIdNot(riskCategoryEntity.getName(), riskCategoryEntity.getId())) {
             throw new DuplicateResourceException("Já existe uma categoria de risco com este nome!");
@@ -64,11 +66,13 @@ public class RiskCategoryService {
         return riskCategoryRepository.save(riskCategoryEntity);
     }
 
+    @Transactional(readOnly = true)
     public RiskCategoryEntity findById(Long id) {
         return riskCategoryRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Categoria de risco não encontrada!"));
     }
 
+    @Transactional(readOnly = true)
     public List<RiskCategoryEntity> findAll() {
         return riskCategoryRepository.findAllByOrderByIdAsc();
     }
